requirements-dev.txt
-r requirements.txt
debugpy==1.8.14
pylint
pylint-django
ipython
